+++ packages/plugin/docs/examples/simple/config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from '@dovenv/core'
 import {
 	getCurrentDir,
 	joinPath,
+	relativePath,
 } from '@dovenv/core/utils'
 
 import { docsPlugin } from '../../src/main'
@@ -10,8 +11,8 @@ const currDir = joinPath( getCurrentDir( import.meta.url ) )
 const pkgDir  = joinPath( currDir, '..', '..' )
 export default defineConfig(
 	docsPlugin( {
-		in   : joinPath( currDir, 'docs' ),
-		out  : joinPath( pkgDir, 'build', 'simple' ),
+		in   : relativePath( process.cwd(), joinPath( currDir, 'docs' ) ),
+		out  : relativePath( process.cwd(), joinPath( pkgDir, 'build', 'simple' ) ),
 		name : 'simple-example',
 
 	} ) )