+++ packages/utils/src/media/icons/main.example.ts
@@ -1,0 +1,7 @@
+import {
+	fontAwesomeRegularIcons,
+	fa2svg,
+} from './main'
+
+const svg = await fa2svg( fontAwesomeRegularIcons.faBell )
+console.log( svg )