+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/core
 
+## 0.0.17
+
+### Patch Changes
+
+- Update plugins (docs, utils, etc)
+
+- Updated dependencies []:
+  - @dovenv/utils@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes