+++ packages/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dovenv
 
+## 0.0.16
+
+### Patch Changes
+
+- Update plugins (docs, utils, etc)
+
 ## 0.0.15
 
 ### Patch Changes