+++ packages/plugin/lint/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @dovenv/lint
 
+## 0.0.17
+
+### Patch Changes
+
+- Update plugins (docs, utils, etc)
+
+- Updated dependencies []:
+  - @dovenv/eslint-config@0.0.16
+  - @dovenv/stylelint-config@0.0.16
+  - @dovenv/core@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes