+++ packages/theme/pigeonposse/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @dovenv/theme-pigeonposse
 
+## 0.0.17
+
+### Patch Changes
+
+- Update plugins (docs, utils, etc)
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.17
+  - @dovenv/theme-banda@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes